+++ test/tests/adapters/mongoose.js
@@ -3,22 +3,19 @@
 
 var expect = require('chai').expect;
 var mongoose = require('mongoose');
-console.log('Mongoose Version: ' + mongoose.version);
-var SchemaAdapter = require('../adapters/mongoose');
+var SchemaAdapter = require('../../../adapters/mongoose');
 
-var isMongoose3 = parseInt(mongoose.version.split('.')[0], 10) === 3;
+describe('Adapters > SchemaAdapter', function () {
+  afterEach(function (done) {
+    delete mongoose.models.Foo;
+    delete mongoose.modelSchemas.Foo;
+    delete mongoose.models.User;
+    delete mongoose.modelSchemas.User;
+    delete mongoose.models.Bar;
+    delete mongoose.modelSchemas.Bar;
+    done();
+  });
 
-afterEach(function (done) {
-  delete mongoose.models.Foo;
-  delete mongoose.modelSchemas.Foo;
-  delete mongoose.models.User;
-  delete mongoose.modelSchemas.User;
-  delete mongoose.models.Bar;
-  delete mongoose.modelSchemas.Bar;
-  done();
-});
-
-describe('SchemaAdapter', function () {
   describe('Date', function () {
     it('should have the type `Date`', function (done) {
       var schema = mongoose.Schema({ foo: Date });
@@ -472,32 +469,30 @@ describe('SchemaAdapter', function () {
     });
   });
 
-  if (!isMongoose3) {
-    describe('Array of schemas with reserved keyword "type"', function () {
-      it('should have the type `{ fields: [...]}`', function (done) {
-        var schema = mongoose.Schema({
-          action: [{
-            type: String,
-            value: String,
-          }]
-        });
-        var model = mongoose.model('Foo', schema);
-
-        return new SchemaAdapter(model, {
-          mongoose: mongoose,
-          connections: [mongoose]
-        })
-          .then(function (schema) {
-            expect(schema.fields[0].type[0].fields).to.be.eql([
-              { field: 'type', type: 'String' },
-              { field: 'value', type: 'String' },
-            ]);
-
-            done(null);
-          });
+  describe('Array of schemas with reserved keyword "type"', function () {
+    it('should have the type `{ fields: [...]}`', function (done) {
+      var schema = mongoose.Schema({
+        action: [{
+          type: String,
+          value: String,
+        }]
       });
+      var model = mongoose.model('Foo', schema);
+
+      return new SchemaAdapter(model, {
+        mongoose: mongoose,
+        connections: [mongoose]
+      })
+        .then(function (schema) {
+          expect(schema.fields[0].type[0].fields).to.be.eql([
+            { field: 'type', type: 'String' },
+            { field: 'value', type: 'String' },
+          ]);
+
+          done(null);
+        });
     });
-  }
+  });
 
   describe('Shower of nested objects/arrays', function () {
     it('should have the type fields: [{...}, {...}]', function (done) {